requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.0.10
django-cors-headers==3.7.0
django-extensions==3.1.3
django-mssql-backend==2.8.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
ldap3==2.9
pyasn1==0.4.8
pycparser==2.20
pyodbc==4.0.30
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
